fix bug in maximizeCash()
crashed when daemon was run with very low ram due to me
neglecting to check for low-ram scenarios. If ram is too low
daemon doesn't even try to maximise server money so this
error was easy to miss. It now handles scenarios where only a
few threads can be used.